feat: refactoring and styled-components setting
