--- week7/hyunn522/src/ProductCategoryRow.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const ProductCategoryRow = ({ category }) => {
-  return (
-    <tr>
-      <th colSpan={2}>{category}</th>
-    </tr>
-  )
-}
-
-export default ProductCategoryRow
