+++ week7/hyunn522/src/components/ProductTable.jsx
@@ -1,0 +1,28 @@
+import React from 'react'
+import ProductRow from './ProductRow';
+
+const ProductTable = ({ products, setProducts, deleteProduct }) => {
+    const productByCategory = products.reduce((acc, product) => {
+        acc[product.category] = acc[product.category] || [];
+        acc[product.category].push(product);
+        return acc;
+    }, {});
+
+
+  return (
+    <div>
+        {Object.keys(productByCategory).map((category) => (
+            <div key={category}>
+                <h3>{category}</h3>
+                <ul>
+                    {productByCategory[category].map((product) => (
+                        <ProductRow key={product.name} product={product} />
+                    ))}
+                </ul>
+            </div>
+        ))}
+    </div>
+  )
+}
+
+export default ProductTable;
